app3/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
python-dateutil==2.8.2
